refactor: change GameManager.kt package directory
