+++ src/main/kotlin/blackjack/controller/BlackJackController.kt
@@ -1,6 +1,6 @@
 package blackjack.controller
 
-import blackjack.model.GameManager
+import blackjack.controller.GameManager
 import blackjack.model.PlayerFactory
 import blackjack.model.Statistics
 import blackjack.view.InputView
